✨A LOT OF CHANGES: ☕изменения в сущностях ☕репозитории к каждой сущности ☕добавление новых сервисов ☕метод addSession в контроллере сессий ☕создание dto session ☕новая папка exceptions: свои удобные исключения для всех контроллеров 🩷изменение post-запроса в timer-settings.js 🩷небольшие изменения index.html ИЗМЕНЕНИЕ СТРУКТУРЫ ПРОЕКТА: ☕удаление Thymeleaf! 🩷перенос фронта в другие папки

+++ src/main/java/com/example/StudyFriends/repositories/SessionRep.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SessionRep extends JpaRepository<Session, Long> {
+public interface SessionRep extends JpaRepository<Session, Integer> {
 
 
 }
